requirements..txt
yt-dlp==2024.10.22
python-telegram-bot==20.8
ffmpeg-python